src/requirements.txt
dotenv>=0.9.9
azure-identity>=1.12.0
microsoft-teams-apps>=2.0.0a5,<3.0.0
O365
pydantic-settings
# === Core Agent Frameworks ===
langchain-core==1.0.2
# langchain-community removed due to conflicts
langgraph==1.0.2
langchain_community
botbuilder-core

# === LLM Providers ===
openai==1.109.1
langchain-openai==1.0.1
google-generativeai==0.3.2
langchain-google-genai==0.0.1

# === Utilities & Tools ===
python-dotenv==1.2.1
requests==2.32.5
python-dateutil==2.9.0.post0
ipython==9.6.0

# === PDF & Reporting Tools ===
pdfkit==1.0.0
reportlab==4.4.4

# === API Server ===
fastapi==0.120.4
uvicorn==0.38.0
faiss-cpu
pandas
sentence_transformers
